Bundle backported TBC versions of all libraries
All of these libraries were extracted from ElvUI-TBC, since they've
already done the hard work of backporting modern versions of Ace3
and ensuring that everything works in TBC.

And thankfully, SyncFriends was written in pure Ace code, and all
WoW API calls and events it uses were supported (and identical)
in TBC too. So this backporting is now complete!