Enhance: Color the Options friends list by action
When you view the SyncFriends addon options, you see a friends list
containing all friends from all of your characters. That's where you
manage all of the data and mark how you want each to be synced (such
as "add to all friends lists" or "remove from all friends lists").

However, the list was extremely bad. Every player name was in pure
white, and it was impossible to know if people were marked for deletion
or addition or skipping. You had to manually click on every entry to
see what they were marked as... It was just HORRIBLE.

To fix that situation, this commit adds colored prefixes to all friends
in the SyncFriends GUI now:

++ (green) = Add.
-- (red) = Remove.
** (gray) = Skip.

Names are formatted with the prefix first, which makes it super easy
to see what's going on... A name such as "++ Greatguy" or "-- Badguy"
instantly tells you all you need to know, without any clicks!