Enhance: Add support for DoIKnowYou player notes
As a user of DoIKnowYou (remembers who people are), it was extremely
frustrating to navigate the SyncFriends list, since it was just a long
list of names without any way to see my reminders about who they are.

That made it very difficult to manage the list and make decisions such
as "well, that person isn't really a close friend anymore, I'll delete
them" or "that person should definitely be synced to all alts".

To see DoIKnowYou notes you had to painfully and manually type the
slash command `/diky <name of person>` to bring up the DoIKnowYou
window... one by one... for every user...

It was torture... so this commit implements deep, built-in integration
with the DoIKnowYou (and DoIKnowYouEnhanced) addon:

* When you hover over players in the SyncFriends list, the tooltip will
  display their "DoIKnowYou: Positive (+2)" or similar rating, as well
  as the player note.

* If you edit the rating or note for someone in DoIKnowYou, your changes
  are immediately detected and updated in the SyncFriends player list.

* Supports the extended features of "DoIKnowYouEnhanced" such as
  "trusted note authors" which lets you see notes written by people
  you trust, if you haven't written anything yourself.

If you're a DoIKnowYou user (or someone interested in that feature),
then you should ABSOLUTELY install the improved "DoIKnowYouEnhancedTBC"
version of that addon, which contains tons of extremely serious (game-
breaking) bug fixes and new features. It is the best version of the best
addon for remembering other players:

https://github.com/VideoPlayerCode/DoIKnowYouEnhancedTBC