Enhance: Output chat warning when add/remove fails
Sometimes, your list contains people who have deleted or renamed their
character, or moved it to the other faction. In that case, the internal
"Friend Queue" system (added in a previous commit) attempts five times
and then gives up. That could be a bit confusing for people, for example
seeing five "Friends must be part of your alliance." WoW error messages
in a row in their chat.

Therefore, this implements a final SyncFriends error message whenever
the "Friend Queue" system gives up due to failing all attempts.

In that case, it now outputs one of the following final messages:

- "SyncFriends: Could not add Theplayer"
- "SyncFriends: Could not remove Theplayer"

This lets the player know who/what failed, so that they can open their
SyncFriends options and delete that person (via the "Forget" button).

PS: SyncFriends itself is already fully faction-aware, and stores your
friend-lists independently for the Horde and Alliance sides, if anyone
is reading this and is curious about that...